+++ package.json
@@ -90,8 +90,8 @@
     "@testing-library/react": "^11.0.4",
     "@testing-library/user-event": "^12.1.5",
     "@types/proj4": "^2.5.0",
-    "@typescript-eslint/eslint-plugin": "^4.9.1",
-    "@typescript-eslint/parser": "^4.9.1",
+    "@typescript-eslint/eslint-plugin": "^4.14.0",
+    "@typescript-eslint/parser": "^4.14.0",
     "eslint-config-airbnb-typescript": "^12.0.0",
     "eslint-import-resolver-typescript": "^2.3.0",
     "eslint-plugin-cypress": "^2.11.2",
