+++ src/common/types/utils.ts
@@ -1,0 +1,2 @@
+export type PartialExcept<T, K extends keyof T> = Partial<T> & Pick<T, K>;
+export type Optional<T, K extends keyof T> = Omit<T, K> & Partial<T>;
