+++ src/domain/hanke/list/testInitialData.ts
@@ -6,6 +6,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_1',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-26T00:00:00Z',
     loppuPvm: '2020-11-17T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -26,6 +27,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_2',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-26T00:00:00Z',
     loppuPvm: '2020-11-27T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -46,6 +48,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_3',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-20T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -66,6 +69,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_4',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-25T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -86,6 +90,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_5',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-23T00:00:00Z',
     loppuPvm: '2020-11-20T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -106,6 +111,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_6',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -126,6 +132,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_7',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -146,6 +153,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_8',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -166,6 +174,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_9',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -186,6 +195,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_10',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-02T00:00:00Z',
     loppuPvm: '2020-12-05T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -206,6 +216,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_11',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-02T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -226,6 +237,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_12',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-02T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -246,6 +258,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_13',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-02T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -266,6 +279,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_14',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-02T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -286,6 +300,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_15',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-24T00:00:00Z',
     loppuPvm: '2020-12-26T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -306,6 +321,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_16',
     onYKTHanke: false,
     nimi: 'j',
+    kuvaus: '',
     alkuPvm: '2020-12-17T00:00:00Z',
     loppuPvm: '2020-12-19T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -326,6 +342,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_35',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-27T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -346,6 +363,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_17',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-18T00:00:00Z',
     loppuPvm: '2021-01-02T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -366,6 +384,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_23',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-02T00:00:00Z',
     loppuPvm: '2020-12-11T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -386,6 +405,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_18',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-24T00:00:00Z',
     loppuPvm: '2020-12-26T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -406,6 +426,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_24',
     onYKTHanke: false,
     nimi: 'dsffff',
+    kuvaus: '',
     alkuPvm: '2020-12-16T00:00:00Z',
     loppuPvm: '2020-12-19T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -426,6 +447,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_19',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-09T00:00:00Z',
     loppuPvm: '2020-12-26T00:00:00Z',
     vaihe: 'SUUNNITTELU',
@@ -446,6 +468,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_20',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-26T00:00:00Z',
     vaihe: 'SUUNNITTELU',
@@ -466,6 +489,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_25',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-10T00:00:00Z',
     loppuPvm: '2020-12-18T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -486,6 +510,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_21',
     onYKTHanke: false,
     nimi: 'dsffff',
+    kuvaus: '',
     alkuPvm: '2020-12-16T00:00:00Z',
     loppuPvm: '2020-12-19T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -506,6 +531,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_36',
     onYKTHanke: false,
     nimi: 'h',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -526,6 +552,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_22',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-30T00:00:00Z',
     loppuPvm: '2021-01-02T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -546,6 +573,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_37',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-23T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -587,6 +615,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_38',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-23T00:00:00Z',
     loppuPvm: '2020-11-27T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -607,6 +636,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_26',
     onYKTHanke: false,
     nimi: 'sfd',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-20T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -627,6 +657,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_27',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-12-22T00:00:00Z',
     loppuPvm: '2020-12-19T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -647,6 +678,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_39',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-27T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -667,6 +699,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_28',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -687,6 +720,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_29',
     onYKTHanke: false,
     nimi: 'cc',
+    kuvaus: '',
     alkuPvm: '2020-12-07T00:00:00Z',
     loppuPvm: '2020-12-12T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -707,6 +741,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_30',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-27T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -727,6 +762,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_31',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-26T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -747,6 +783,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_32',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-26T00:00:00Z',
     loppuPvm: '2020-11-17T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -767,6 +804,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_33',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -787,6 +825,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_34',
     onYKTHanke: true,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-28T00:00:00Z',
     vaihe: 'OHJELMOINTI',
@@ -807,6 +846,7 @@ export const initialData: HankeDataDraft[] = [
     hankeTunnus: 'SMTGEN2_40',
     onYKTHanke: false,
     nimi: 'dsf',
+    kuvaus: '',
     alkuPvm: '2020-11-24T00:00:00Z',
     loppuPvm: '2020-11-27T00:00:00Z',
     vaihe: 'OHJELMOINTI',
